Improved tagging logic
